Refactor files structure

+++ routes/posts.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const db = require('../data/helpers/posts');
+const { body, param, validationResult } = require('express-validator/check');
 
 router.get('/', async (req, res) => {
   // TODO: use query params for search
@@ -10,7 +11,6 @@ router.get('/', async (req, res) => {
 
 router.get('/:id', async (req, res) => {
   const [post] = await db.getPost(req.params.id);
-
   if (!post) {
     return res.status(404).json({ error: 'Post not Found' });
   }
@@ -19,9 +19,8 @@ router.get('/:id', async (req, res) => {
 });
 
 router.get('/:postId/comments', async (req, res) => {
+  // validates post id first
   const [post] = await db.getPost(req.params.postId);
-
-  // validates post id
   if (!post) {
     return res.status(404).json({ error: "Post doesn't exist" });
   }
@@ -30,31 +29,53 @@ router.get('/:postId/comments', async (req, res) => {
   return res.json(comments);
 });
 
-router.post('/', async (req, res) => {
-  // TODO: validate fields
-  const newPost = {
-    title: req.body.title,
-    url: req.body.url,
-    authorId: 1 // TODO: get author by AUTH params
-  };
+// prettier-ignore
+router.post('/', [
+    body('title')
+      .isLength({ min: 5, max: 128 }).trim().escape()
+      .withMessage('You must include a post title.'),
+    body('url')
+      .isURL().trim().escape()
+      .withMessage('You must include a valid URL.')
+  ],
+  async (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(422).json({ errors: errors.array() });
+    }
 
-  const [id] = await db.addPost(newPost);
-  const post = await db.getPost(id);
+    const newPost = {
+      title: req.body.title,
+      url: req.body.url,
+      authorId: 1 // TODO: get author by AUTH params
+    };
 
-  return res.status(201).json(post);
-});
+    const [id] = await db.addPost(newPost);
+    const post = await db.getPost(id);
 
-router.delete('/:postId', async (req, res) => {
-  // TODO: validate author with AUTH id
-  const post = await db.deletePost(req.params.postId);
-
-  if (!post) {
-    return res.status(404).json({ error: "Post doesn't exist" });
+    return res.status(201).json(post);
   }
+);
 
-  return res.status(204).json();
-});
+// prettier-ignore
+router.delete('/:postId', [
+    param('postId').isInt().withMessage('Invalid ID.')
+  ],
+  async (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res.status(422).json({ errors: errors.array() });
+    }
 
-// updates a post
+    // TODO: validate author with AUTH id
+    const post = await db.deletePost(req.params.postId);
+
+    if (!post) {
+      return res.status(404).json({ error: "Post doesn't exist" });
+    }
+
+    return res.status(204).json();
+  }
+);
 
 module.exports = router;
